Refactor widget.component to widgets.component Change widget title font-size
