+++ src/app/root-content/dashboard/widgets/widgets-menu/widgets-menu.component.ts
@@ -1,6 +1,6 @@
 import { Component, Input, Output, EventEmitter } from "@angular/core";
 
-import { Widget } from "../widget.component";
+import { Widget } from "../widgets.component";
 
 import * as $ from "jquery";
