+++ src/app/root-content/dashboard/widgets/widgets.component.scss
@@ -6,6 +6,7 @@
 
 	.ui.cards {
 		.ui.card {
+			font-size: .95em;
 			width: 245px;
 
 			/deep/ .widget-header {
